+++ terra_opp/serializers.py
@@ -3,6 +3,7 @@ from typing import Optional
 from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.core.exceptions import ObjectDoesNotExist
+from geostore import GeometryTypes
 from geostore.models import Feature, Layer
 from rest_framework import serializers
 from rest_framework.fields import SerializerMethodField
@@ -143,11 +144,15 @@ class ViewpointSerializerWithPicture(serializers.ModelSerializer):
         related_docs = validated_data.pop('related', None)
         point_data = validated_data.pop('point', None)
         layer, created = Layer.objects.get_or_create(
-            name=settings.TROPP_BASE_LAYER_NAME
+            pk=settings.TROPP_OBSERVATORY_LAYER_PK,
+            defaults={
+                'geom_type': GeometryTypes.Point,
+                'id': settings.TROPP_OBSERVATORY_LAYER_PK
+            }
         )
         feature = Feature.objects.create(
             geom=point_data.get('geom'),
-            layer=layer,
+            layer_id=settings.TROPP_OBSERVATORY_LAYER_PK,
             properties={},
         )
         validated_data.setdefault('point', feature)